+++ android/src/main/kotlin/com/gtbluesky/fusion/container/FusionActivity.kt
@@ -65,10 +65,6 @@ open class FusionActivity : FlutterActivity(), FusionContainer {
         engineBinding = null
     }
 
-    override fun onBackPressed() {
-        engineBinding?.pop(true)
-    }
-
     override fun provideFlutterEngine(context: Context) = engineBinding?.engine
 
     override fun providePlatformPlugin(